feat: Order controller, global handler and mapper with tests

+++ domain/domain-core/src/main/java/com/tui/proof/domain/valueobject/Quantity.java
@@ -6,5 +6,5 @@ import lombok.Data;
 @Builder
 @Data
 public class Quantity {
-    private final Integer value;
+    private final Long value;
 }
